requirements.txt
numpy~=1.23.5
torch~=1.12.0
Pillow~=9.4.0
SimpleITK~=2.2.1
torchvision~=0.13.0
einops~=0.6.1
timm~=0.6.13
scipy~=1.10.1
h5py~=3.9.0
prettytable~=3.7.0
MedPy~=0.4.0
packaging~=23.1
opencv-python~=4.7.0.72
wandb~=0.14.2
matplotlib~=3.7.1
rich~=13.4.2
monai~=1.3.0
dynamic_network_architectures~=0.2
nibabel~=5.1.0
mmcv~=2.0.1
mmengine~=0.8.4
mmpretrain~=1.0.2
scikit-learn~=1.2.2
mmrazor~=1.0.0
argparse~=1.4.0
mmdet~=3.1.0
synapseclient~=3.1.0
cupy-cuda11x
cucim